refactor(home): move WavyDivider to common/ui
