+++ apps/frontend/src/common/ui/WavyDivider.vue
